refactor(ast): dedupe shared scanner primitives into ast_signal_utils
Consolidate the byte-identical _line_number_from_index helper (copied
across ast_csharp, ast_go, ast_java, ast_php, ast_ruby, ast_rust, and
ast_swift) and the _balanced_segment brace/paren matcher (copied across
ast_csharp, ast_java, ast_ruby, ast_rust) into ast_signal_utils, and
import them from there.

ast_go keeps a local _balanced_segment override because Go has backtick
raw string literals, in which backslashes are literal; unifying it to
the common variant would miscount brace/paren depth inside raw strings.
Behavior is preserved for every language.
