+++ src/agent_bom/ast_signal_utils.py
@@ -1,8 +1,41 @@
-"""Shared prompt and guardrail signal helpers for AST analyzers."""
+"""Shared prompt, guardrail, and source-scanning helpers for AST analyzers."""
 
 from __future__ import annotations
 
 import re
+
+
+def _line_number_from_index(source: str, index: int) -> int:
+    return source[:index].count("\n") + 1
+
+
+def _balanced_segment(source: str, open_index: int, *, open_char: str, close_char: str) -> tuple[str, int] | None:
+    if open_index < 0 or open_index >= len(source) or source[open_index] != open_char:
+        return None
+    depth = 0
+    in_quote = ""
+    escaped = False
+    for index in range(open_index, len(source)):
+        char = source[index]
+        if in_quote:
+            if char == "\\" and not escaped:
+                escaped = True
+                continue
+            if char == in_quote and not escaped:
+                in_quote = ""
+            escaped = False
+            continue
+        if char in {'"', "'"}:
+            in_quote = char
+            escaped = False
+            continue
+        if char == open_char:
+            depth += 1
+        elif char == close_char:
+            depth -= 1
+            if depth == 0:
+                return source[open_index : index + 1], index + 1
+    return None
 
 _GUARDRAIL_CALL_PATTERNS = re.compile(
     r"\b(?:content_filter|safety_check|moderate|moderation|validate_input|"
